requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
Django==5.2.7
django-allauth==65.12.0
django-countries==7.6.1
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
dotenv==0.9.9
idna==3.10
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
stripe==13.0.0
typing_extensions==4.15.0
urllib3==2.5.0
